backend/requirements.txt
asgiref==3.10.0
diff-match-patch==20241021
Django==5.2.8
django-cors-headers==4.9.0
django-import-export==4.3.14
djangorestframework==3.16.1
sqlparse==0.5.3
tablib==3.9.0
whitenoise==6.11.0
